Changed CIT_PATH_PETSC to support the new config layout in petsc-dev ('bmake' -> 'conf').  The macro still recognizes the old layout. Also, it now expects a full three-component version number as its first argument -- i.e., in 'configure.ac', instead of
    CIT_PATH_PETSC([2.3])

one must now write

    CIT_PATH_PETSC([2.3.3])

Because -- as it turns out -- releases of PETSc that differ only in
the subminor version can be incompatible with one another.

It is unclear to me what logic they are using.  Version 2.3.3 is not
merely a bug-fix release; it is the first version to include Sieve,
and contains changes to existing function prototypes/signatures.  If
that doesn't justify an increment to v2.4.x, what does?